Show icon for autoplay in control panel and modify test app (#197)
* Show icon for autoplay in control panel and modify test app

* Remove commented method
