+++ preview/MsixCore/Tests/notepadplus.msix
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:bc838717529809a5ac5eefed213205b89e7c51b1ae03cdb0b23670c87c24888a
-size 3112115
+oid sha256:7bb0c5a3f3c3e720c1b5091975bd5b8d4ccf2a834e5dd5b8771ab0a6f35dc642
+size 3112113
